// src/utils/errors/types.ts
/**
 * Types of errors that can occur in the application
 */
export enum ErrorType {
  // Authentication errors
  AUTH_EMAIL_NOT_CONFIRMED = "AUTH_EMAIL_NOT_CONFIRMED",
  AUTH_INVALID_CREDENTIALS = "AUTH_INVALID_CREDENTIALS",
  AUTH_USER_NOT_FOUND = "AUTH_USER_NOT_FOUND",
  AUTH_EMAIL_IN_USE = "AUTH_EMAIL_IN_USE",
  AUTH_WEAK_PASSWORD = "AUTH_WEAK_PASSWORD",
  AUTH_INVALID_RESET_TOKEN = "AUTH_INVALID_RESET_TOKEN",
  
  // Profile errors
  PROFILE_NOT_FOUND = "PROFILE_NOT_FOUND",
  PROFILE_UPDATE_FAILED = "PROFILE_UPDATE_FAILED",
  
  // Merchant errors
  MERCHANT_NOT_FOUND = "MERCHANT_NOT_FOUND",
  MERCHANT_UPDATE_FAILED = "MERCHANT_UPDATE_FAILED",
  MERCHANT_INVALID_DATA = "MERCHANT_INVALID_DATA",
  
  // Stamp card errors
  STAMP_CARD_NOT_FOUND = "STAMP_CARD_NOT_FOUND",
  STAMP_CARD_CREATE_FAILED = "STAMP_CARD_CREATE_FAILED",
  STAMP_CARD_UPDATE_FAILED = "STAMP_CARD_UPDATE_FAILED",
  
  // QR code errors
  QR_CODE_INVALID = "QR_CODE_INVALID",
  QR_CODE_EXPIRED = "QR_CODE_EXPIRED",
  QR_CODE_ALREADY_USED = "QR_CODE_ALREADY_USED",
  QR_CODE_GENERATION_FAILED = "QR_CODE_GENERATION_FAILED",
  
  // Data operations errors
  DATA_FETCH_FAILED = "DATA_FETCH_FAILED",
  DATA_CREATE_FAILED = "DATA_CREATE_FAILED",
  DATA_UPDATE_FAILED = "DATA_UPDATE_FAILED",
  DATA_DELETE_FAILED = "DATA_DELETE_FAILED",
  
  // Stamp operation errors
  STAMP_ISSUE_FAILED = "STAMP_ISSUE_FAILED",
  STAMP_REDEEM_FAILED = "STAMP_REDEEM_FAILED",
  
  // General errors
  NETWORK_ERROR = "NETWORK_ERROR",
  UNKNOWN_ERROR = "UNKNOWN_ERROR",
  VALIDATION_ERROR = "VALIDATION_ERROR",
  PERMISSION_DENIED = "PERMISSION_DENIED"
}